Refactor code-compiler to multiple files so it's easier to add support for wasm. Also setup the test for solana, substrate with the new compile endpoint for SO file
