+++ code-compiler/compiler/wasm.py
@@ -1,0 +1,111 @@
+import urllib.parse
+import os
+import subprocess
+import threading
+import ipfshttpclient
+import requests
+from distutils.dir_util import copy_tree
+
+from helper.helper import write_to_disk
+
+
+class WasmCodegenAndBuild(threading.Thread):
+    """
+    NpmBuild is class that will run `npm run codegen && npm run build` in a new thread, not blocking the main thread
+
+    """
+
+    def __init__(self, generated_folder):
+        self.stdout = None
+        self.stderr = None
+        self.generated_folder = generated_folder
+        threading.Thread.__init__(self)
+
+    def run(self):
+        try:
+            output = subprocess.check_output(["npm install && npm run codegen && npm run build"],
+                                             stderr=subprocess.STDOUT,
+                                             shell=True, universal_newlines=True, cwd=self.generated_folder)
+        except subprocess.CalledProcessError as exc:
+            print("Compilation has failed. The result can be found in: " + self.generated_folder)
+            write_to_disk(self.generated_folder + "/error.txt", exc.output)
+        else:
+            print("Compilation was success. The result can be found in: " + self.generated_folder)
+            write_to_disk(self.generated_folder + "/success.txt", output)
+
+
+def compile_wasm(data, hash):
+    # Create new folder
+    generated_folder = "generated/" + hash
+    os.mkdir(generated_folder)
+    os.mkdir(generated_folder + "/src")
+    os.mkdir(generated_folder + "/abis")
+
+    # URL-decode the data
+    mapping = urllib.parse.unquote_plus(data["mapping.ts"])  # We currently support uploading 1 mapping file only
+    subgraph = urllib.parse.unquote_plus(data["subgraph.yaml"])
+    schema = urllib.parse.unquote_plus(data["schema.graphql"])
+    abis = urllib.parse.unquote_plus(data["abis"])
+    project_name = urllib.parse.unquote_plus(data["project_name"])
+
+    # Populating stub data
+    copy_tree("stub-wasm", os.path.join(generated_folder))
+
+    # Save the formatted data from request to disk, ready for compiling
+    write_to_disk(generated_folder + "/src/mapping.ts", mapping)
+    write_to_disk(generated_folder + "/subgraph.yaml", subgraph)
+    write_to_disk(generated_folder + "/schema.graphql", schema)
+    write_to_disk(generated_folder + "/abis/" + project_name + ".json", abis)
+
+    # Codegen & Build
+    print("Generating code + compiling for: " + hash + ". This will take a while!")
+    wasm_codegen_and_build = WasmCodegenAndBuild(generated_folder)
+    wasm_codegen_and_build.start()
+
+    return hash
+
+
+def deploy_wasm(data):
+    # Parse the data
+    compilation_id = urllib.parse.unquote_plus(data["compilation_id"])
+    project_name = urllib.parse.unquote_plus(data["project_name"])
+
+    # Get the files path from generated/hash folder
+    project = os.path.join("./generated", compilation_id, "subgraph.yaml")
+    mapping = os.path.join("./generated", compilation_id, "build", project_name, project_name + ".wasm")
+    schema = os.path.join("./generated", compilation_id, "schema.graphql")
+
+    # Uploading files to IPFS
+    if os.environ.get('IPFS_URL'):
+        client = ipfshttpclient.connect(os.environ.get('IPFS_URL'))  # Connect with IPFS container name
+    else:
+        client = ipfshttpclient.connect()
+
+    print("Uploading files to IPFS...")
+    config_res = client.add(project)
+    mapping_res = client.add(mapping)
+    schema_res = client.add(schema)
+
+    # Uploading to IPFS result
+    print("project.yaml: " + config_res['Hash'])
+    print(project_name + ".wasm: " + mapping_res['Hash'])
+    print("schema.graphql: " + schema_res['Hash'])
+
+    # Uploading IPFS files to Index Manager
+    if os.environ.get('INDEX_MANAGER_URL'):
+        index_manager_url = os.environ.get('INDEX_MANAGER_URL')  # Connection to indexer
+    else:
+        index_manager_url = 'http://0.0.0.0:3030'
+
+    res = requests.post(index_manager_url,
+                        json={
+                            'jsonrpc': '2.0',
+                            'method': 'index_deploy',
+                            'params': [
+                                config_res['Hash'],
+                                mapping_res['Hash'],
+                                schema_res['Hash']
+                            ],
+                            'id': 1,
+                        })
+    print(res.json())
